First raw version of refactor
