+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/AbstractCoordinator.java
@@ -19,7 +19,6 @@ package org.apache.kafka.clients.consumer.internals;
 import org.apache.kafka.clients.ClientResponse;
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.Node;
-import org.apache.kafka.common.errors.AuthenticationException;
 import org.apache.kafka.common.errors.DisconnectException;
 import org.apache.kafka.common.errors.GroupAuthorizationException;
 import org.apache.kafka.common.errors.IllegalGenerationException;
@@ -48,7 +47,6 @@ import org.apache.kafka.common.requests.LeaveGroupResponse;
 import org.apache.kafka.common.requests.OffsetCommitRequest;
 import org.apache.kafka.common.requests.SyncGroupRequest;
 import org.apache.kafka.common.requests.SyncGroupResponse;
-import org.apache.kafka.common.utils.KafkaThread;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.common.utils.Time;
 import org.slf4j.Logger;
@@ -59,7 +57,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * AbstractCoordinator implements group management for a single group member by interacting with
@@ -91,9 +88,11 @@ import java.util.concurrent.atomic.AtomicReference;
  * when sending a request that affects the state of the group (e.g. JoinGroup, LeaveGroup).
  */
 public abstract class AbstractCoordinator implements Closeable {
+    // TODO remove this
     public static final String HEARTBEAT_THREAD_PREFIX = "kafka-coordinator-heartbeat-thread";
+    private final LogContext logContext;
 
-    private enum MemberState {
+    enum MemberState {
         UNJOINED,    // the client is not part of a group
         REBALANCING, // the client has begun rebalancing
         STABLE,      // the client has joined and is sending heartbeats
@@ -134,6 +133,7 @@ public abstract class AbstractCoordinator implements Closeable {
                                Time time,
                                long retryBackoffMs,
                                boolean leaveGroupOnClose) {
+        this.logContext = logContext;
         this.log = logContext.logger(AbstractCoordinator.class);
         this.client = client;
         this.time = time;
@@ -203,6 +203,14 @@ public abstract class AbstractCoordinator implements Closeable {
     public synchronized void ensureCoordinatorReady() {
         // Using zero as current time since timeout is effectively infinite
         ensureCoordinatorReady(0, Long.MAX_VALUE);
+    }
+
+    MemberState state() {
+        return state;
+    }
+
+    RequestFuture<Void> findCoordinatorFuture() {
+        return findCoordinatorFuture;
     }
 
     /**
@@ -318,7 +326,12 @@ public abstract class AbstractCoordinator implements Closeable {
 
     private synchronized void startHeartbeatThreadIfNeeded() {
         if (heartbeatThread == null) {
-            heartbeatThread = new HeartbeatThread();
+            heartbeatThread = new HeartbeatThread(logContext,
+                    groupId,
+                    heartbeat,
+                    this,
+                    time,
+            retryBackoffMs);
             heartbeatThread.start();
         }
     }
@@ -757,6 +770,7 @@ public abstract class AbstractCoordinator implements Closeable {
     }
 
     // visible for testing
+    // TODO consider moving to HeartbeatThread
     synchronized RequestFuture<Void> sendHeartbeatRequest() {
         log.debug("Sending Heartbeat request to coordinator {}", coordinator);
         HeartbeatRequest.Builder requestBuilder =
@@ -882,142 +896,6 @@ public abstract class AbstractCoordinator implements Closeable {
                 "The number of seconds since the last controller heartbeat was sent"),
                 lastHeartbeat);
         }
-    }
-
-    private class HeartbeatThread extends KafkaThread {
-        private boolean enabled = false;
-        private boolean closed = false;
-        private AtomicReference<RuntimeException> failed = new AtomicReference<>(null);
-
-        private HeartbeatThread() {
-            super(HEARTBEAT_THREAD_PREFIX + (groupId.isEmpty() ? "" : " | " + groupId), true);
-        }
-
-        public void enable() {
-            synchronized (AbstractCoordinator.this) {
-                log.debug("Enabling heartbeat thread");
-                this.enabled = true;
-                heartbeat.resetTimeouts(time.milliseconds());
-                AbstractCoordinator.this.notify();
-            }
-        }
-
-        public void disable() {
-            synchronized (AbstractCoordinator.this) {
-                log.debug("Disabling heartbeat thread");
-                this.enabled = false;
-            }
-        }
-
-        public void close() {
-            synchronized (AbstractCoordinator.this) {
-                this.closed = true;
-                AbstractCoordinator.this.notify();
-            }
-        }
-
-        private boolean hasFailed() {
-            return failed.get() != null;
-        }
-
-        private RuntimeException failureCause() {
-            return failed.get();
-        }
-
-        @Override
-        public void run() {
-            try {
-                log.debug("Heartbeat thread started");
-                while (true) {
-                    synchronized (AbstractCoordinator.this) {
-                        if (closed)
-                            return;
-
-                        if (!enabled) {
-                            AbstractCoordinator.this.wait();
-                            continue;
-                        }
-
-                        if (state != MemberState.STABLE) {
-                            // the group is not stable (perhaps because we left the group or because the coordinator
-                            // kicked us out), so disable heartbeats and wait for the main thread to rejoin.
-                            disable();
-                            continue;
-                        }
-
-                        client.pollNoWakeup();
-                        long now = time.milliseconds();
-
-                        if (coordinatorUnknown()) {
-                            if (findCoordinatorFuture != null || lookupCoordinator().failed())
-                                // the immediate future check ensures that we backoff properly in the case that no
-                                // brokers are available to connect to.
-                                AbstractCoordinator.this.wait(retryBackoffMs);
-                        } else if (heartbeat.sessionTimeoutExpired(now)) {
-                            // the session timeout has expired without seeing a successful heartbeat, so we should
-                            // probably make sure the coordinator is still healthy.
-                            markCoordinatorUnknown();
-                        } else if (heartbeat.pollTimeoutExpired(now)) {
-                            // the poll timeout has expired, which means that the foreground thread has stalled
-                            // in between calls to poll(), so we explicitly leave the group.
-                            maybeLeaveGroup();
-                        } else if (!heartbeat.shouldHeartbeat(now)) {
-                            // poll again after waiting for the retry backoff in case the heartbeat failed or the
-                            // coordinator disconnected
-                            AbstractCoordinator.this.wait(retryBackoffMs);
-                        } else {
-                            heartbeat.sentHeartbeat(now);
-
-                            sendHeartbeatRequest().addListener(new RequestFutureListener<Void>() {
-                                @Override
-                                public void onSuccess(Void value) {
-                                    synchronized (AbstractCoordinator.this) {
-                                        heartbeat.receiveHeartbeat(time.milliseconds());
-                                    }
-                                }
-
-                                @Override
-                                public void onFailure(RuntimeException e) {
-                                    synchronized (AbstractCoordinator.this) {
-                                        if (e instanceof RebalanceInProgressException) {
-                                            // it is valid to continue heartbeating while the group is rebalancing. This
-                                            // ensures that the coordinator keeps the member in the group for as long
-                                            // as the duration of the rebalance timeout. If we stop sending heartbeats,
-                                            // however, then the session timeout may expire before we can rejoin.
-                                            heartbeat.receiveHeartbeat(time.milliseconds());
-                                        } else {
-                                            heartbeat.failHeartbeat();
-
-                                            // wake up the thread if it's sleeping to reschedule the heartbeat
-                                            AbstractCoordinator.this.notify();
-                                        }
-                                    }
-                                }
-                            });
-                        }
-                    }
-                }
-            } catch (AuthenticationException e) {
-                log.error("An authentication error occurred in the heartbeat thread", e);
-                this.failed.set(e);
-            } catch (GroupAuthorizationException e) {
-                log.error("A group authorization error occurred in the heartbeat thread", e);
-                this.failed.set(e);
-            } catch (InterruptedException | InterruptException e) {
-                Thread.interrupted();
-                log.error("Unexpected interrupt received in heartbeat thread", e);
-                this.failed.set(new RuntimeException(e));
-            } catch (Throwable e) {
-                log.error("Heartbeat thread failed due to unexpected error", e);
-                if (e instanceof RuntimeException)
-                    this.failed.set((RuntimeException) e);
-                else
-                    this.failed.set(new RuntimeException(e));
-            } finally {
-                log.debug("Heartbeat thread has closed");
-            }
-        }
-
     }
 
     protected static class Generation {
